+++ lib/src/extensions.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+import 'theme_manager.dart';
+
+extension ThemeColorExtension on String {
+  Color get tc => themeFusion.getColor(this);
+}