+++ lib/src/theme_colors_base.dart
@@ -1,8 +1,0 @@
-import 'package:flutter/material.dart';
-
-abstract class BaseThemeColors {
-  const BaseThemeColors();
-  Color get primary;
-  Color get background;
-  Color get text;
-}