+++ lib/theme_fusion.dart
@@ -1,5 +1,5 @@
 library theme_fusion;
 
 export 'src/theme_fusion_app.dart';
-export 'src/theme_fusion_core.dart';
-export 'src/theme_colors_base.dart';
+export 'src/theme_manager.dart';
+export 'src/extensions.dart';